+++ CurrencyConversionApp/Stores/CurrencyConverterStore.swift
@@ -8,8 +8,8 @@ final class CurrencyCoverterStore {
     let currencies: Observable<[Currency]>
     private let _currencies = BehaviorRelay<[Currency]>.init(value: [])
 
-    let usdRates: Observable<[String: Float]>
-    private let _usdRates = BehaviorRelay<[String: Float]>.init(value: [:])
+    let usdRates: Observable<[String: Double]>
+    private let _usdRates = BehaviorRelay<[String: Double]>.init(value: [:])
     
     let disposeBag = DisposeBag()
